requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.47.1
langchain==0.3.27
langchain-google-genai==2.0.10
google-generativeai==0.8.5
chromadb==1.0.15
yt-dlp==2025.7.21
moviepy==1.0.3
opencv-python==4.12.0.88
openai-whisper==20250625
ffmpeg==1.4
pydantic==2.11.7
python-dotenv==1.1.1
python-slugify==8.0.4
redis==6.2.0